requirements.txt
apscheduler
asgiref==3.3.1
backcall==0.2.0
black
celery
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
decorator==4.4.2
defusedxml==0.7.1
Django>=3
django-templated-mail==1.1.1
djangorestframework>=3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
flake8
idna==2.10
ipdb==0.13.7
ipython==7.21.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.17
pycparser==2.20
Pygments==2.8.1
psycopg2-binary
PyJWT
python3-openid==3.2.0
pytz==2021.1
redis
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
